fix(common): removed Wno-format flag and fixed formatting warnings (IDFGH-7768)

+++ components/esp_modem/examples/modem_tcp_client/main/sock_dce.hpp
@@ -128,7 +128,7 @@ esp_modem::return_type name(__VA_ARGS__);
             return false;
         }
         if (state != status::IDLE) {
-            ESP_LOGE("dce", "Unexpected state %d", state);
+            ESP_LOGE("dce", "Unexpected state %d", static_cast<int>(state));
             return false;
         }
         return true;
